+++ inst/shiny/app.R
@@ -37,7 +37,7 @@ ui <- page_navbar(
 
 server <- function(input, output, session) {
   rlang::check_installed(
-    c("promises", "shinychat", "ellmer", "mirai"),
+    c("promises", "shinychat", "ellmer", "later", "mirai"),
     reason = "to run the tempest app."
   )
   if (!"chat_server" %in% getNamespaceExports("shinychat")) {